requirements.txt
Flask>=1.1.2
Flask-Cors>=3.0.8
kafka-python>=2.0.1
python-dotenv>=0.14.0
